requirements.txt
--find-links https://storage.googleapis.com/jax-releases/jax_cuda_releases.html

jax[cuda12_pip]
flax
matplotlib==3.7.5
scipy==1.12.0
numpy
scikit-learn
opencv-python
statsmodels
seaborn
ipython
ipykernel
imageio
hydra-core
wandb
grpcio
protobuf==3.20.*
